updated to rev 7609
updated to rev 7633

updated to rev 7654

updating upto rev 7801

updated to rev 7938

updated to rev 7945

updated to rev 7968

updated to rev 7999

Updated to rev 8419

updated to rev 8592

updated to rev 8604

updated to rev 8648

updated to rev 8693

updated to rev 8740

updated to rev 9816

updated to rev 9882

updated to rev 9882

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/caseaggregation/CaseAggregationConditionService.java
@@ -61,9 +61,8 @@ public interface CaseAggregationConditionService
     CaseAggregationCondition getCaseAggregationCondition( DataElement dataElement, DataElementCategoryOptionCombo optionCombo );
     
     Collection<CaseAggregationCondition> getCaseAggregationCondition( Collection<DataElement> dataElements );
-
     
-    Double parseConditition( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
+    Integer parseConditition( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
         
     Collection<PatientDataValue> getPatientDataValues( CaseAggregationCondition aggregationCondition, OrganisationUnit orgunit, Period period );
